# perf(filter): reduce allocations in parsing metric