# Reduce allocations in parsing metric

Reduced the number of allocations of the `restoreMetricStringByNameAndLabels` function by setting the capacity in `strings.Builder`, now the function makes stable 2 allocations to `strings.Builder` and to slice creation

```
benchmark                                          old ns/op     new ns/op     delta
BenchmarkRestoreMetricStringByNameAndLabels-12     1071          851           -20.57%

benchmark                                          old allocs     new allocs     delta
BenchmarkRestoreMetricStringByNameAndLabels-12     6              2              -66.67%

benchmark                                          old bytes     new bytes     delta
BenchmarkRestoreMetricStringByNameAndLabels-12     624           320           -48.72%
```